# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API reto Torneo Solidario",
        "description": "API reto Torneo Solidario",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Obtener todos los usuarios",
                "description": "Obtener todos los usuarios",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/users"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Crear un usuario",
                "description": "Crear un nuevo usuario",
                "operationId": "postUser",
                "requestBody": {
                    "description": "Datos del usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "activo",
                                    "perfil"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Usuario 1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": "1"
                                    },
                                    "perfil": {
                                        "type": "string",
                                        "example": "administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no válidos"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Mostrar usuario",
                "description": "Mostrar un usuario por su id",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario mostrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Actualizar un usuario",
                "description": "Actualizar un usuario por su id",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "activo",
                                    "perfil"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Usuario1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": "1"
                                    },
                                    "perfil": {
                                        "type": "string",
                                        "example": "administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Eliminar un usuario",
                "description": "Eliminar un usuario por su id",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "users": {
                "title": "users",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Usuario 1"
                    },
                    "email": {
                        "type": "string",
                        "example": "usuario@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": "1"
                    },
                    "perfil": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        }
    },
...
}